backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
gitpython==3.1.43
radon==6.0.1
vulture==2.11
pathspec==0.12.1
python-multipart==0.0.9
